chore: reorg vfs file
